+++ src/command_center/widgets/__init__.py
@@ -1,0 +1,2 @@
+"""Command Center widget helpers for ms-markets integrations."""
+